refactoring user controller and services and adding exceptions

+++ src/main/java/com/sysmap/parrot/application/requests/User/UserService.java
@@ -5,15 +5,18 @@ import com.sysmap.parrot.application.requests.User.GetUser.GetUserResponse;
 import com.sysmap.parrot.application.requests.User.UpdateUser.UpdateUserRequest;
 import com.sysmap.parrot.application.requests.User.UpdateUser.UpdateUserResponse;
 import com.sysmap.parrot.application.requests.fileUpload.IFileUploadService;
+import com.sysmap.parrot.domain.entities.Connection;
 import com.sysmap.parrot.domain.entities.User;
+import com.sysmap.parrot.exception.badRequestException.BadRequestException;
+import com.sysmap.parrot.exception.notFoundException.NotFoundException;
 import com.sysmap.parrot.infrastructure.data.IUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -30,6 +33,11 @@ public class UserService implements IUserService {
 
     public String createUser(CreateUserRequest request) {
 
+        var registeredUser = _userRepository.getUserByEmail(request.email);
+
+        if (registeredUser.isPresent()) {
+            throw new BadRequestException("This email is already in use. Try another one!");
+        }
         var encodedPassword = _passwordEncoder.encode(request.password);
 
         var user = new User(request.name, request.email);
@@ -41,24 +49,33 @@ public class UserService implements IUserService {
     }
 
     public GetUserResponse getUserById(String id) {
-        var uuid = UUID.fromString(id);
-        var user = _userRepository.findById(uuid).get();
+        var userById = _userRepository.findById(UUID.fromString(id));
 
-        var response = new GetUserResponse(user.getId(), user.getName(), user.getEmail());
+        if (userById.isEmpty()) {
+            throw new NotFoundException("There is no user with this id!");
+        }
+
+        var user = userById.get();
+
+        var response = new GetUserResponse(user.getId(), user.getName(), user.getEmail(), user.getPictureUrl());
 
         return response;
     }
 
     public UpdateUserResponse updateUserById(String id, UpdateUserRequest request) {
-        var uuid = UUID.fromString(id);
-        var user = _userRepository.findById(uuid).get();
+        var userById = _userRepository.findById(UUID.fromString(id));
+
+        if(userById.isEmpty()) {
+            throw new NotFoundException("There is no user with this id!");
+        }
+
+        var user = userById.get();
 
         var encodedPassword = _passwordEncoder.encode(request.password);
 
         user.setName(request.name);
         user.setEmail(request.email);
         user.setPassword(encodedPassword);
-        user.setPictureUrl(request.pictureUrl);
 
         _userRepository.save(user);
 
@@ -66,37 +83,58 @@ public class UserService implements IUserService {
                 user.getId(),
                 user.getName(),
                 user.getEmail(),
-                user.getPassword(),
-                user.getPictureUrl()
+                user.getPassword()
         );
 
         return response;
     }
 
     public String deleteUserById(String id) {
-        var uuid = UUID.fromString(id);
-        var user = _userRepository.findById(uuid).get();
+        var userById = _userRepository.findById(UUID.fromString(id));
+
+        if(userById.isEmpty()) {
+            throw new NotFoundException("There is no user with this id!");
+        }
+
+        var user = userById.get();
 
         _userRepository.delete(user);
 
-        var response = "Usuário excluído com sucesso";
-        return response;
+        return "Usuário excluído com sucesso";
     }
 
-    public void uploadPhoto(MultipartFile photo) throws Exception {
-        var user = ((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal());
+    public void uploadPhoto(String id, MultipartFile photo) {
+        var userById = _userRepository.findById(UUID.fromString(id));
+
+        if(userById.isEmpty()) {
+            throw new NotFoundException("There is no user with this id!");
+        }
+
+        var user = userById.get();
 
         var photoUri = "";
 
         try {
             var fileName = user.getId() + "." + photo.getOriginalFilename().substring(photo.getOriginalFilename().lastIndexOf(".") + 1);
             photoUri = _fileUploadService.upload(photo, fileName);
-        } catch (Exception e) {
-            throw new Exception(e.getMessage());
+        } catch (BadRequestException e) {
+            throw new BadRequestException("Unable to upload image!");
         }
 
         user.setPictureUrl(photoUri);
         _userRepository.save(user);
 
+    }
+
+    public User getUser(String email) {
+        return _userRepository.getUserByEmail(email).get();
+    }
+
+    public User findUserById(UUID id) {
+        return _userRepository.findById(id).get();
+    }
+
+    public List<User> getAllUsers() {
+        return _userRepository.findAll();
     }
 }
